Use `script` & `script_env` to install `certifi`
Make the build & installation process of `certifi` OS agnostic using
`meta.yaml` and Jinja to smooth over the differences. This is already
`noarch: python` and builds only happen on Linux. So the Windows
`bld.bat` is not necessarily staying up-to-date anyways. Better to
consolidate into one process flow that works for both seamlessly.